Create ComponentLocator. Create components folder. Create base folder for locators and exceptions. Create GuidBasePage. Create shells of support, user, and search pages.
