+++ base/exceptions.py
