+++ pages/base.py
@@ -1,135 +1,20 @@
 import settings
 
-from selenium.webdriver.common.by import By
-from selenium.webdriver.support.ui import WebDriverWait
-from selenium.webdriver.support import expected_conditions as EC
+import urllib.parse
+from selenium.common.exceptions import NoSuchElementException
 from selenium.webdriver.common.action_chains import ActionChains
-from selenium.common.exceptions import StaleElementReferenceException, TimeoutException, NoSuchElementException
+from selenium.webdriver.common.by import By
 
-from pages.exceptions import HttpError, PageException
-
-
-class WebElementWrapper:
-
-    def __init__(self, driver, element_name, locator):
-        self.driver = driver
-        self.locator = locator
-        self.name = element_name
-
-    def __getattr__(self, item):
-        return getattr(self.element, item)
-
-    @property
-    def element(self):
-        return self.locator.get_web_element(self.driver, self.name)
-
-    def present(self):
-        try:
-            self.element
-            return True
-        except ValueError:
-            return False
-
-    def absent(self, timeout=settings.QUICK_TIMEOUT):
-        """
-        Boolean to check if an element is no longer visible on a page.
-        """
-        try:
-            WebDriverWait(self.driver, timeout).until(
-                EC.invisibility_of_element_located(self.locator.location)
-            )
-            return True
-        except TimeoutException:
-            return False
-
-    def here_then_gone(self):
-        self.present()  # Allow a wait for it to appear, but don't force it to have been present
-        if not self.absent():
-            raise ValueError('Element {} is not absent.'.format(self.name))
-        return True
-
-
-class BaseLocator:
-
-    def __init__(self, selector, path, timeout=settings.TIMEOUT):
-        self.selector = selector
-        self.path = path
-        self.location = (selector, path)
-        self.timeout = timeout
-
-
-class Locator(BaseLocator):
-
-    def get_web_element(self, driver, element):
-        """
-        Checks if element is on page, visible, and clickable before returning the selenium webElement.
-
-        This method is adapted from code provided on seleniumframework.com
-        """
-        try:
-            WebDriverWait(driver, self.timeout).until(
-                EC.presence_of_element_located(self.location)
-            )
-        except(TimeoutException, StaleElementReferenceException):
-            raise ValueError('Element {} not present on page. {}'.format(element, driver.current_url)) from None
-
-        try:
-            WebDriverWait(driver, self.timeout).until(
-                EC.visibility_of_element_located(self.location)
-            )
-        except(TimeoutException, StaleElementReferenceException):
-            raise ValueError('Element {} not visible before timeout. {}'.format(element, driver.current_url)) from None
-
-        if 'link' in element:
-            try:
-                WebDriverWait(driver, self.timeout).until(
-                    EC.element_to_be_clickable(self.location)
-                )
-            except(TimeoutException, StaleElementReferenceException):
-                raise ValueError('Element {} on page but not clickable. {}'.format(element, driver.current_url)) from None
-
-        return driver.find_element(self.selector, self.path)
-
-
-class GroupLocator(BaseLocator):
-    def get_web_elements(self, driver):
-        return driver.find_elements(self.selector, self.path)
-
-
-class BaseElement:
-    default_timeout = settings.TIMEOUT
-
-    def __new__(cls, *args, **kwargs):
-        page = None
-        if hasattr(cls, 'waffle_override'):
-            for waffle_name in cls.waffle_override:
-                if waffle_name in settings.EMBER_PAGES:
-                    page = super().__new__(cls.waffle_override[waffle_name])
-        else:
-            page = super().__new__(cls)
-        page.__init__(*args, **kwargs)  # TODO: is there a way to not have to do this?
-        return page
-
-    def __init__(self, driver):
-        self.driver = driver
-
-    def verify(self):
-        raise NotImplementedError
-
-    def __getattribute__(self, item):
-        value = object.__getattribute__(self, item)
-        if type(value) is Locator:
-            return WebElementWrapper(self.driver, item, value)
-        elif type(value) is GroupLocator:
-            return value.get_web_elements(self.driver)
-        return value
+from components.navbars import HomeNavbar
+from base.locators import BaseElement, ComponentLocator
+from base.exceptions import HttpError, PageException
 
 
 class BasePage(BaseElement):
     url = None
 
     def __init__(self, driver, verify=False):
-        super(BasePage, self).__init__(driver)
+        super().__init__(driver)
         if verify:
             self.check_page()
 
@@ -163,43 +48,13 @@ class BasePage(BaseElement):
         # Note: If you close the browser too quickly, the drag/drop may not go through
 
 
-class Navbar(BaseElement):
-    service_dropdown = Locator(By.CSS_SELECTOR, '#navbarScope > div.container > div.navbar-header > div.dropdown > button')
-    home_link = Locator(By.CSS_SELECTOR, '#navbarScope > div.container > div.navbar-header > div.dropdown > ul > li:nth-child(1) > a')
-    preprints_link = Locator(By.CSS_SELECTOR, '#navbarScope > div.container > div.navbar-header > div.dropdown > ul > li:nth-child(2) > a')
-    registries_link = Locator(By.CSS_SELECTOR, '#navbarScope > div.container > div.navbar-header > div.dropdown > ul > li:nth-child(3) > a')
-    meetings_link = Locator(By.CSS_SELECTOR, '#navbarScope > div.container > div.navbar-header > div.dropdown > ul > li:nth-child(4) > a')
-    search_link = Locator(By.ID, 'navbar-search')
-    support_link = Locator(By.ID, 'navbar-support')
-    donate_link = Locator(By.ID, 'navbar-donate')
-    user_dropdown = Locator(By.CSS_SELECTOR, 'a.dropdown-toggle')
-    user_dropdown_profile = Locator(By.CSS_SELECTOR, '#secondary-navigation > ul > li.dropdown.open > ul > li:nth-child(1) > a')
-    user_dropdown_support = Locator(By.CSS_SELECTOR, '#secondary-navigation > ul > li.dropdown.open > ul > li:nth-child(2) > a')
-    user_dropdown_settings = Locator(By.CSS_SELECTOR, '#secondary-navigation > ul > li.dropdown.open > ul > li:nth-child(3) > a')
-    sign_up_button = Locator(By.LINK_TEXT, 'Sign Up')
-    sign_in_button = Locator(By.LINK_TEXT, 'Sign In')
-    logout_link = Locator(By.CSS_SELECTOR, '#secondary-navigation > ul > li.dropdown.open > ul > li:nth-child(4) > a')
-    current_service = Locator(By.CSS_SELECTOR, '#navbarScope > div > div > div.service-home > a > span.current-service > strong')
-
-    def verify(self):
-        return len(self.driver.find_elements(By.ID, 'navbarScope')) == 1
-
-    def is_logged_in(self):
-        return self.user_dropdown.present()
-
-    def is_logged_out(self):
-        return self.sign_in_button.present()
-
-
 class OSFBasePage(BasePage):
     url = settings.OSF_HOME
-
+    navbar = ComponentLocator(HomeNavbar)
     # all pages must have a unique identity or overwrite verify
 
     def __init__(self, driver, verify=False):
-        super(OSFBasePage, self).__init__(driver)
-
-        self.navbar = self.BasePageNavbar(driver)
+        super().__init__(driver)
 
         if verify:
             self.check_page()
@@ -227,10 +82,8 @@ class OSFBasePage(BasePage):
     def is_logged_out(self):
         return self.navbar.is_logged_out()
 
-    class BasePageNavbar(Navbar):
 
-        # Locators
-        my_project_link = Locator(By.CSS_SELECTOR, '#secondary-navigation > ul > li:nth-last-child(5) > a')
-
-        def verify(self):
-            return self.current_service.text == 'HOME'
+class GuidBasePage(OSFBasePage):
+    def __init__(self, driver, verify=False, guid=None):
+        super().__init__(driver, verify)
+        self.url = urllib.parse.urljoin(settings.OSF_HOME, guid) if guid else None
