+++ pages/landing.py
@@ -1,11 +1,20 @@
 import settings
 
-from pages.base import OSFBasePage, Locator
 from selenium.webdriver.common.by import By
+
+from pages.base import OSFBasePage
+from components.navbars import EmberNavbar
+from base.locators import Locator, ComponentLocator
+
+
+class EmberLandingPage(OSFBasePage):
+    identity = Locator(By.ID, 'home-hero', settings.LONG_TIMEOUT)
+
+    # Components
+    navbar = ComponentLocator(EmberNavbar)
 
 
 class LandingPage(OSFBasePage):
-    url = settings.OSF_HOME
+    waffle_override = {'ember_home_page': EmberLandingPage}
 
-    # Locators
     identity = Locator(By.ID, 'home-hero', settings.LONG_TIMEOUT)
