+++ pages/project.py
@@ -1,16 +1,17 @@
 import settings
-import urllib.parse
 
 from selenium.webdriver.common.by import By
-from pages.base import OSFBasePage, Locator
+
+from base.locators import Locator
+from pages.base import GuidBasePage, OSFBasePage
 
 
-class ProjectPage(OSFBasePage):
-
-    # Locators
+class ProjectPage(GuidBasePage):
     identity = Locator(By.CSS_SELECTOR, '#overview > nav#projectSubnav')
     project_title = Locator(By.ID, 'nodeTitleEditable')
 
-    def __init__(self, driver, verify=False, guid=None):
-        super(self.__class__, self).__init__(driver, verify)
-        self.url = urllib.parse.urljoin(settings.OSF_HOME, guid) if guid else None
+
+class MyProjectsPage(OSFBasePage):
+    url = settings.OSF_HOME + '/myprojects/'
+
+    identity = Locator(By.CSS_SELECTOR, '.col-xs-8 > h3:nth-child(1)')
