+++ pages/support.py
@@ -1,0 +1,18 @@
+import settings
+
+from selenium.webdriver.common.by import By
+
+from base.locators import Locator
+from pages.base import OSFBasePage
+
+
+class EmberSupportPage(OSFBasePage):
+    url = settings.OSF_HOME + '/support'
+
+    identity = Locator(By.CLASS_NAME, '__bd682', settings.LONG_TIMEOUT)
+
+
+class SupportPage(OSFBasePage):
+    waffle_override = {'ember_support_page': EmberSupportPage}
+
+    identity = Locator(By.CSS_SELECTOR, 'body > div.watermarked > div > div > h1', settings.LONG_TIMEOUT)
